Add idx field to Interface
Knowing the interface index (also known as the interface number) is required for
setting some socket options, particularly around IPv6 multicasting (e.g.
[UdpSocket#join_multicast_v6](https://doc.rust-lang.org/std/net/struct.UdpSocket.html#method.join_multicast_v6)).

This adds an idx field on Interface (an `Option<u32>`) and populates it on posix
systems. Unfortunately, I am not a Windows programmer, so I'm leaving it
unimplemented there; I would be delighted to see someone else pick up the torch
and get it implemented on Windows as well.