Improve support for niche platforms
This adds support for visionOS and `target_env = "newlib"` (embedded) platforms by always providing an implementation, such that only one of these `cfg` lists have to be maintained.

A little bit of context on why these names are not just unified in `libc` can be found in https://github.com/rust-lang/libc/issues/682.